🌅[TAWS-61] Updated source to apply webpack config
