--- source/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<title>Armadillo</title>
-	<link rel="icon" type="image/png" href="./brand/favicon/favicon.ico">
-	<link rel="stylesheet" type="text/css" href="./styles/style.css">
-	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.13.0/css/all.min.css">
-</head>
-
-<body>
-	<nav>
-		<ul class="menu">
-			<li class="logo"><a href="./index.html">Armadillo</a></li>
-			<li class="item"><a href="./index.html">Home</a></li>
-			<li class="item"><a href="./pages/projects/projects.html">Projects</a></li>
-			<li class="item"><a href="./pages/about-us/about-us-page.html">About</a></li>
-			<li class="toggle"><a href="#"><i class="fas fa-bars"></i></a></li>
-		</ul>
-	</nav>
-	<h1 style="display: flex; justify-content: center;">Welcome to Armadillo!</h1>
-	<script src="./scripts/app.js"></script>
-</body>
-
-</html>
