--- source/pages/about-us/about-us.html
@@ -1,80 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<title>About Us</title>
-	<link rel="stylesheet" type="text/css" href="./styles/about-us.css">
-	<link rel="stylesheet" type="text/css" href="../../styles/style.css">
-	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.13.0/css/all.min.css">
-</head>
-
-<body>
-	<nav>
-		<ul class="menu">
-			<li class="logo"><a href="/source/index.html">Armadillo</a></li>
-			<li class="item"><a href="/source/index.html">Home</a></li>
-			<li class="item"><a href="../projects/projects.html">Projects</a></li>
-			<li class="item"><a href="./about-us-page.html">About</a></li>
-			<li class="toggle"><a href="#"><i class="fas fa-bars"></i></a></li>
-		</ul>
-	</nav>
-	<div class="center">
-		<h1>Yep, this is the team behind <a href="/source/index.html">Armadillo</a></h1>
-	</div>
-	<div class="team-cards">
-		<div class="card">
-			<img src="../../images/minion.jpg" alt="Avatar" style="width:100%">
-			<div class="card-text">
-				<h4><b>Annerlee</b></h4>
-				<p>Design & Development</p>
-			</div>
-		</div>
-		<div class="card">
-			<img src="../../images/minion.jpg" alt="Avatar" style="width:100%">
-			<div class="card-text">
-				<h4><b>Brandon</b></h4>
-				<p>Design & Development</p>
-			</div>
-		</div>
-		<div class="card">
-			<img src="../../images/minion.jpg" alt="Avatar" style="width:100%">
-			<div class="card-text">
-				<h4><b>Marius</b></h4>
-				<p>Lead Development</p>
-			</div>
-		</div>
-		<div class="card">
-			<img src="../../images/minion.jpg" alt="Avatar" style="width:100%">
-			<div class="card-text">
-				<h4><b>Matt</b></h4>
-				<p>Lead Development</p>
-			</div>
-		</div>
-		<div class="card">
-			<img src="../../images/minion.jpg" alt="Avatar" style="width:100%">
-			<div class="card-text">
-				<h4><b>Ojo</b></h4>
-				<p>Development</p>
-			</div>
-		</div>
-		<div class="card">
-			<img src="../../images/minion.jpg" alt="Avatar" style="width:100%">
-			<div class="card-text">
-				<h4><b>Sarah</b></h4>
-				<p>Design & Development</p>
-			</div>
-		</div>
-		<div class="card">
-			<img src="../../images/minion.jpg" alt="Avatar" style="width:100%">
-			<div class="card-text">
-				<h4><b>Zach</b></h4>
-				<p>Project Management & Development</p>
-			</div>
-		</div>
-	</div>
-	<script src="../../scripts/app.js"></script>
-</body>
-
-</html>
