--- source/pages/contact-us/contact-us.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<link rel="stylesheet" href="../../styles/style.css">
-	<link rel="stylesheet" href="./styles/contact-us.css">
-	<title>Contact Us</title>
-</head>
-<body>
-	<div class="container">
-		<form action="">
-			<div class="section">
-				<label for="name" id="name">Name</label>
-				<input class="item" type="name" id="name" placeholder="Please enter your name">
-			</div>
-			<div class="section">
-				<label for="email" id="email">Email</label>
-				<input class="item" type="email" id="email" placeholder="Email address">
-			</div>
-			<div class="section">
-				<label for="request" id="request">How may we help you?</label>
-				<textarea class="item" name="request" id="request" cols="30" rows="10" placeholder="Please enter your question here..."></textarea>
-			</div>
-			<div class="section">
-				<button class="item button btn-primary" type="submit" id="SubmitBtn" style="width:80px;height:40px;">Submit</button>
-			</div>
-		</form>
-	</div>
-</body>
-</html>
