--- source/pages/projects/projects.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<link rel="stylesheet" type="text/css" href="../../styles/style.css">
-	<link rel="stylesheet" type="text/css" href="./styles/projects.css">
-	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.13.0/css/all.min.css">
-	<title>Projects</title>
-</head>
-
-<body>
-	<nav>
-		<ul class="menu">
-			<li class="logo"><a href="/source/index.html">Armadillo</a></li>
-			<li class="item"><a href="/source/index.html">Home</a></li>
-			<li class="item"><a href="./pages/projects/projects.html">Projects</a></li>
-			<li class="item"><a href="../about-us/about-us-page.html">About</a></li>
-			<li class="toggle"><a href="#"><i class="fas fa-bars"></i></a></li>
-		</ul>
-	</nav>
-	<div id="wrapper" class="y">
-		<section>
-			<div class="front tile">
-				<div class="mobile">
-					<img src="../../brand/favicon/favicon.ico" alt="Website">
-				</div>
-			</div>
-			<div class="back tile">
-				<p>Team Armadillo Website - or short TAWS</p>
-			</div>
-		</section>
-		<section>
-			<div class="front tile">
-				<img src="../images/dashboard.png" alt="Codepen">
-			</div>
-			<div class="back tile">
-				<p>Armadillo Team Management App - or short ATMA</p>
-			</div>
-		</section>
-	</div>
-	<script src="../../scripts/app.js"></script>
-</body>
-</html>
