+++ source/scripts/index.js
@@ -1,0 +1,3 @@
+import "@/styles/index.css";
+
+import "@/scripts/components/navbar.js";
