+++ source/scripts/pages/about-us.js
@@ -1,0 +1,8 @@
+import "@/styles/pages/about-us.css";
+import "@/images/minion.jpg";
+
+import template from "@/templates/pages/about-us.html";
+
+window.addEventListener("DOMContentLoaded", function () {
+	document.body.insertAdjacentHTML("beforeend", template);
+});
