+++ source/scripts/pages/contact-us.js
@@ -1,0 +1,7 @@
+import "@/styles/pages/contact-us.css";
+
+import template from "@/templates/pages/contact-us.html";
+
+window.addEventListener("DOMContentLoaded", function () {
+	document.body.insertAdjacentHTML("beforeend", template);
+});
