+++ source/scripts/pages/projects.js
@@ -1,0 +1,9 @@
+import "@/styles/pages/projects.css";
+import "@/brand/favicon/favicon.ico";
+import "@/images/dashboard.png";
+
+import template from "@/templates/pages/projects.html";
+
+window.addEventListener("DOMContentLoaded", function () {
+	document.body.insertAdjacentHTML("beforeend", template);
+});
