+++ source/styles/components/navbar.css
@@ -1,13 +1,3 @@
-/* Basic styling */
-* {
-	box-sizing: border-box;
-	padding: 0;
-	margin: 0;
-}
-body {
-	font-family: Arial, Helvetica, sans-serif;
-	font-size: 16px;
-}
 nav {
 	background: antiquewhite;
 	padding: 0 15px;
