+++ source/styles/index.css
@@ -1,0 +1,10 @@
+/* Basic styling */
+* {
+	box-sizing: border-box;
+	padding: 0;
+	margin: 0;
+}
+body {
+	font-family: Arial, Helvetica, sans-serif;
+	font-size: 16px;
+}
