+++ source/templates/components/navbar.html
@@ -1,0 +1,13 @@
+<nav>
+	<ul class="menu">
+		<li class="logo"><a href="./index.html">Armadillo</a></li>
+		<li class="item"><a href="./index.html">Home</a></li>
+		<li class="item"><a href="./projects.html">Projects</a></li>
+		<li class="item"><a href="./about-us.html">About</a></li>
+		<li class="item"><a href="./contact-us.html">Contact Us</a></li>
+
+		<li class="toggle">
+			<a href="#"><i class="fas fa-bars"></i></a>
+		</li>
+	</ul>
+</nav>
