+++ source/templates/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<meta charset="UTF-8" />
+		<title><%= htmlWebpackPlugin.options.title %></title>
+		<link rel="icon" href="/assets/brand/favicon/favicon.ico" />
+		<link
+			rel="stylesheet"
+			href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.13.0/css/all.min.css"
+		/>
+	</head>
+
+	<body></body>
+</html>
