+++ source/templates/pages/about-us.html
@@ -1,0 +1,85 @@
+<div class="center">
+	<h1>
+		Yep, this is the team behind
+		<a href="./index.html">Armadillo</a>
+	</h1>
+</div>
+<div class="team-cards">
+	<div class="card">
+		<img
+			src="/assets/images/minion.jpg"
+			alt="Avatar"
+			style="width: 100%;"
+		/>
+		<div class="card-text">
+			<h4><b>Annerlee</b></h4>
+			<p>Design & Development</p>
+		</div>
+	</div>
+	<div class="card">
+		<img
+			src="/assets/images/minion.jpg"
+			alt="Avatar"
+			style="width: 100%;"
+		/>
+		<div class="card-text">
+			<h4><b>Brandon</b></h4>
+			<p>Design & Development</p>
+		</div>
+	</div>
+	<div class="card">
+		<img
+			src="/assets/images/minion.jpg"
+			alt="Avatar"
+			style="width: 100%;"
+		/>
+		<div class="card-text">
+			<h4><b>Marius</b></h4>
+			<p>Lead Development</p>
+		</div>
+	</div>
+	<div class="card">
+		<img
+			src="/assets/images/minion.jpg"
+			alt="Avatar"
+			style="width: 100%;"
+		/>
+		<div class="card-text">
+			<h4><b>Matt</b></h4>
+			<p>Lead Development</p>
+		</div>
+	</div>
+	<div class="card">
+		<img
+			src="/assets/images/minion.jpg"
+			alt="Avatar"
+			style="width: 100%;"
+		/>
+		<div class="card-text">
+			<h4><b>Ojo</b></h4>
+			<p>Development</p>
+		</div>
+	</div>
+	<div class="card">
+		<img
+			src="/assets/images/minion.jpg"
+			alt="Avatar"
+			style="width: 100%;"
+		/>
+		<div class="card-text">
+			<h4><b>Sarah</b></h4>
+			<p>Design & Development</p>
+		</div>
+	</div>
+	<div class="card">
+		<img
+			src="/assets/images/minion.jpg"
+			alt="Avatar"
+			style="width: 100%;"
+		/>
+		<div class="card-text">
+			<h4><b>Zach</b></h4>
+			<p>Project Management & Development</p>
+		</div>
+	</div>
+</div>
