+++ source/templates/pages/projects.html
@@ -1,0 +1,20 @@
+<div id="wrapper" class="y">
+	<section>
+		<div class="front tile">
+			<div class="mobile">
+				<img src="/assets/brand/favicon/favicon.ico" alt="Website" />
+			</div>
+		</div>
+		<div class="back tile">
+			<p>Team Armadillo Website - or short TAWS</p>
+		</div>
+	</section>
+	<section>
+		<div class="front tile">
+			<img src="/assets/images/dashboard.png" alt="Codepen" />
+		</div>
+		<div class="back tile">
+			<p>Armadillo Team Management App - or short ATMA</p>
+		</div>
+	</section>
+</div>
